add 4 basic operations in Util.java

+++ app/src/main/java/test/com/calculator/Util.java
@@ -3,26 +3,26 @@ package test.com.calculator;
 
 public class Util {
 
-    public long add (long num1 , long num2){
-        long sum = 0 ;
+    public double add (double num1 , double num2){
+        double sum = 0 ;
         sum = num1 + num2 ;
         return sum;
     }
 
-    public long subtract (long num1 , long num2){
-        long sub = 0;
+    public double subtract (double num1 , double num2){
+        double sub = 0;
         sub = num1 - num2 ;
         return sub ;
     }
 
-    public long multiply (long num1 , long num2){
-        long mul = 0;
+    public double multiply (double num1 , double num2){
+        double mul = 0;
         mul = num1 * num2 ;
         return mul ;
     }
 
-    public long devide (long num1 , long num2){
-        long dev = 0;
+    public double devide (double num1 , double num2){
+        double dev = 0;
         if (num2 !=0){
             dev = num1/num2;
         }
